+++ .github/Changelog.md
@@ -6,3 +6,9 @@
 
 ### Changelog BudgeTeer:
 - Save balances, budgets and open transactions on lock
+- Check budget entry for validity
+- Move balances to header
+- ToDo:
+    - Fix Date picker for budget validity
+    - Check budget validity calculating remaining budget for the current month in the header
+    - Show save button on modification/check/deletion of balances, budgets and open transactions